requirements.txt
amqp==2.6.1
asgiref==3.3.1
Babel==2.9.0
billiard==3.6.3.0
braintree==3.45.0
brew==0.1.3
cairocffi==1.2.0
CairoSVG==2.5.1
celery==4.4.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.1.6
django-crispy-forms==1.11.1
django-tinymce==3.2.0
django-widget-tweaks==1.4.8
flower==0.9.2
html5lib==1.1
idna==2.10
kombu==4.6.11
pdfrw==0.4
Pillow==8.1.0
prompt-toolkit==3.0.14
pycparser==2.20
Pyphen==0.10.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
tornado==4.2
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
WeasyPrint==0.42.3
webencodings==0.5.1
